requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
black==25.1.0
cffi==1.17.1
cfgv==3.4.0
click==8.1.8
cryptography==44.0.2
distlib==0.3.9
ecdsa==0.19.1
fastapi==0.115.12
filelock==3.18.0
greenlet==3.2.0
h11==0.14.0
httptools==0.6.4
identify==2.6.9
idna==3.10
iniconfig==2.1.0
mypy==1.15.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
pluggy==1.5.0
pre_commit==4.2.0
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pytest==8.3.5
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.1
uvloop==0.21.0
virtualenv==20.30.0
watchfiles==1.0.5
websockets==15.0.1
